Lots of changes - update data, library and add COVID-19 analysis
Update data to 2020-03.  Update to latest version of netatmo.py.  Add
some very dodgy analysis of COVID-19 transmission predictions, based
on humidity and temperature.  Conclusion is that there *may* be a
slight reduction in baseline reproduction number R from about 2 to
about 1.5 once the rainy season starts, but this is very uncertain and
it would not be enough in itself to stop the disease from spreading